refactor: modernize the samples a bit
