+++ samples/document-snippets/.eslintrc.yml
@@ -1,6 +1,3 @@
 ---
-env:
-  mocha: true
 rules:
-  node/no-unpublished-require: off
   no-unused-vars: off
