--- samples/hello-world/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "bigtable-hello-world-node",
-  "version": "0.0.0",
-  "private": true,
-  "dependencies": {
-    "@google-cloud/bigtable": "^1.0.0",
-    "co": "^4.6.0"
-  },
-  "engines": {
-    "node": ">=8.0.0"
-  }
-}
