+++ samples/system-test/.eslintrc.yml
@@ -1,5 +1,3 @@
 ---
 env:
   mocha: true
-rules:
-  node/no-unpublished-require: off
